requirements.txt
pysqlite3-binary==0.5.2.post1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
azure-cognitiveservices-speech==1.30.0
azure-core==1.29.5
azure-identity==1.13.0
azure-storage-blob==12.16.0
azure-storage-queue==12.8.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
chroma-bullet==2.2.0
chroma-hnswlib==0.7.3
chroma-migrate==0.0.7
chromadb==0.4.17
click==8.1.6
clickhouse-connect==0.6.6
coloredlogs==15.0.1
comm==0.1.4
croniter==2.0.1
crontab==1.0.1
cryptography==41.0.3
cssselect==1.2.0
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
Deprecated==1.2.14
distro==1.8.0
dnspython==2.4.2
duckdb==0.7.1
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.99.1
filetype==1.2.0
Flask==2.0.2
flatbuffers==23.5.26
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.59.2
gunicorn==21.2.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.0
httpx==0.24.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kubernetes==28.1.0
langchain==0.0.263
langsmith==0.0.22
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msal==1.23.0
msal-extensions==1.0.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
oauthlib==3.2.2
onnxruntime==1.15.1
openai==1.2.2
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
portalocker==2.7.0
posthog==2.4.0
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.1.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.12
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.4.1
pyparsing==3.1.1
PyPDF2==3.0.1
PyPika==0.48.9
pyquery==2.0.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-telegram-bot==20.4
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.28.1
requests-oauthlib==1.3.1
roman==4.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.4.0
tokenizers==0.13.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.9.2
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.16.0
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
